+++ src/backend/joanie/tests/core/test_api_enrollment.py
@@ -143,6 +143,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                                 .isoformat()
                                 .replace("+00:00", "Z"),
                             },
+                            "languages": enrollment.course_run.languages,
                         },
                         "created_on": enrollment.created_on.isoformat().replace(
                             "+00:00", "Z"
@@ -208,6 +209,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                                 .isoformat()
                                 .replace("+00:00", "Z"),
                             },
+                            "languages": other_enrollment.course_run.languages,
                         },
                         "created_on": other_enrollment.created_on.isoformat().replace(
                             "+00:00", "Z"
@@ -277,6 +279,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                                 .isoformat()
                                 .replace("+00:00", "Z"),
                             },
+                            "languages": course_run_1.languages,
                         },
                         "created_on": enrollment_1.created_on.isoformat().replace(
                             "+00:00", "Z"
@@ -424,6 +427,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                         .isoformat()
                         .replace("+00:00", "Z"),
                     },
+                    "languages": enrollment.course_run.languages,
                 },
                 "created_on": enrollment.created_on.isoformat().replace("+00:00", "Z"),
                 "is_active": enrollment.is_active,
@@ -522,6 +526,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                         .isoformat()
                         .replace("+00:00", "Z"),
                     },
+                    "languages": course_run.languages,
                 },
                 "created_on": enrollment.created_on.isoformat().replace("+00:00", "Z"),
                 "is_active": is_active,
@@ -640,6 +645,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                         .isoformat()
                         .replace("+00:00", "Z"),
                     },
+                    "languages": course_run.languages,
                 },
                 "created_on": enrollment.created_on.isoformat().replace("+00:00", "Z"),
                 "is_active": is_active,
@@ -720,6 +726,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                         .isoformat()
                         .replace("+00:00", "Z"),
                     },
+                    "languages": course_run.languages,
                 },
                 "created_on": enrollment.created_on.isoformat().replace("+00:00", "Z"),
                 "is_active": is_active,
@@ -960,6 +967,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                         .isoformat()
                         .replace("+00:00", "Z"),
                     },
+                    "languages": course_run.languages,
                 },
                 "created_on": enrollment.created_on.isoformat().replace("+00:00", "Z"),
                 "is_active": is_active,
@@ -1216,6 +1224,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                             .isoformat()
                             .replace("+00:00", "Z"),
                         },
+                        "languages": enrollment.course_run.languages,
                     },
                     "created_on": enrollment.created_on.isoformat().replace(
                         "+00:00", "Z"
@@ -1413,6 +1422,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                         .isoformat()
                         .replace("+00:00", "Z"),
                     },
+                    "languages": course_run.languages,
                 },
                 "created_on": enrollment.created_on.isoformat().replace("+00:00", "Z"),
                 "is_active": False,
@@ -1482,6 +1492,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                         .isoformat()
                         .replace("+00:00", "Z"),
                     },
+                    "languages": course_run.languages,
                 },
                 "created_on": enrollment.created_on.isoformat().replace("+00:00", "Z"),
                 "is_active": False,
@@ -1538,6 +1549,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                         .isoformat()
                         .replace("+00:00", "Z"),
                     },
+                    "languages": course_run.languages,
                 },
                 "created_on": enrollment.created_on.isoformat().replace("+00:00", "Z"),
                 "is_active": True,
@@ -1616,6 +1628,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                         .isoformat()
                         .replace("+00:00", "Z"),
                     },
+                    "languages": course_run.languages,
                 },
                 "created_on": enrollment.created_on.isoformat().replace("+00:00", "Z"),
                 "is_active": True,
@@ -1672,6 +1685,7 @@ class EnrollmentApiTest(BaseAPITestCase):
                         .isoformat()
                         .replace("+00:00", "Z"),
                     },
+                    "languages": course_run.languages,
                 },
                 "created_on": enrollment.created_on.isoformat().replace("+00:00", "Z"),
                 "is_active": True,